+++ CHANGELOG.md
@@ -7,6 +7,13 @@
 - Cron job for CSV bulk trades
 - Get total invested value per stock
 
+## [0.0.2] - 2024-05-15
+### Added
+- Django Rest Framework
+- Django OAuth Toolkit
+- Django CORS Headers
+- Stock API views
+
 ## [0.0.1] - 2024-05-15
 ### Added
 - Stock Trading `Core` module